new framwork

+++ src/device/function/cuda/embedding_cuda.cu
@@ -3,13 +3,13 @@
 #include "common.h"
 
 
-__global__ void embedding_cuda_kernel(float* y, const float* x, const float* W, const int d, const int x_size) {
+__global__ void embedding_cuda_kernel(float* y, const int* x, const float* W, const int d, const int x_size) {
     // 计算当前线程处理的 token 索引
     int idx = blockIdx.x * blockDim.x + threadIdx.x;
 
     if (idx < x_size) {
         // 获取当前 token 的索引 (输入 token)
-        int token_idx = static_cast<int>(x[idx]);
+        int token_idx = x[idx];
 
         // 每个 token 对应的 embedding 向量
         const float* W_row = W + token_idx * d;
@@ -22,7 +22,7 @@ __global__ void embedding_cuda_kernel(float* y, const float* x, const float* W, 
     }
 }
 
-void embedding_cuda(float* y, const float* x, const float* W, const int d, const int x_size) {
+void embedding_cuda(float* y, const int* x, const float* W, const int d, const int x_size) {
     // 定义线程块和网格的维度
     int block_size = 256;
     int grid_size = (x_size + block_size - 1) / block_size;
